更改协议模式，改为http拉的模式，降低CPU占用

+++ 03_Abstract（组件接口）/FSS.Abstract/Entity/MetaInfo/TaskGroupVO.cs
@@ -65,7 +65,7 @@ namespace FSS.Abstract.Entity.MetaInfo
         /// <summary>
         /// 运行平均耗时
         /// </summary>
-        public int RunSpeedAvg { get; set; }
+        public long RunSpeedAvg { get; set; }
         
         /// <summary>
         /// 运行次数
